Bulk-revert changes caused by AOSP cherry-picks.
In order to bring AOSP development back in-line with master development,
some CLs were cherry-picked individually from internal master to AOSP,
which were then merged back into internal master (MERGED-IN was missing).
Due to merge-conflict pain, these are being reverted in favor of one
big diff.  This CL reverts the changes that were auto-merged in as a result,
and can be used as the target of MERGED-IN when reverting the individual
cherry-picks in AOSP.

This reverts commit a08fe91ee5d162a8376cf4dbe65daed8e0fafcd6, reversing
changes made to 11481d1d950c0681cb3812ead9457ea72876dc4c.

This reverts commit 7ec5ecfbb7f88cc7d71d49ec12333ac78056fa81, reversing
changes made to 6fecbbb27e774487a5b9e4cecd1d735a75345951.

Bug: 37916906
Test: Builds 'n' boots.