Update patch set 4
Patch Set 4:

tad has the following entrypointpaths: ['/res(/.*)?', '/lib(/.*)?', '/sbin(/.*)?', '/init\\..*', '/fstab\\..*', '/ueventd\\..*', '/', '/d', '/etc', '/proc', '/root', '/sdcard', '/config', '/charger', '/bugreports', '/verity_key', '/lost\\+found', '/build\\.prop', '/default\\.prop', '/selinux_version', '/sbin/tad_static']

Because of the MatchPathPrefix(path, "/init") forSystem is set to true. I have now clue why it is assigned '/init\\..*'. However this means that the domain is required to be a 'coredomain'. If it is coredomain services from vendor are not allowed to connect.

As it is located in /sbin it doesn't make sense to mark it as /system.

Patch-set: 4