Update patch set 2
Patch Set 2:

I won't say this with certainty, but is it possible this change has a rather deeper effect on the SE Linux rules than had been intended?  In one device repository a program which had not been granted the dac_override capability was performing actions it could only perform if it /did/ have the dac_override capability.

Might it be that this makes things much easier because it pretty well guts the SE Linux rules?  Alternatively another change might have badly broken SE Linux.

Patch-set: 2
CC: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>